refactor(ui): extract legacy assistant grouping helpers
Slice 11b.1 step 5/14 of the message-part.tsx split. Moves the
legacy assistant-pipeline helpers — list / latestDefined / same /
sameRef / sameGroup / sameGroups / PartRef / PartGroup / groupParts
(legacy) / index / renderable / partDefaultOpen / HIDDEN_TOOLS /
CONTEXT_GROUP_TOOLS / isContextGroupTool — into a dedicated module
`message-part-render-groups.ts`.

These are deliberately kept separate from the v2 grouping helper at
`./message-part-group.ts` (both files own a `groupParts` symbol; the
shapes differ — legacy emits context-tool groups, v2 emits prose /
reasoning / trow-block groups). Inside message-part.tsx the legacy
import is aliased to `legacyGroupParts` so the call sites stay
explicit about which grouping they consume.

The `message-part-stale.test.ts` source-grep that pinned
`latestDefined` to `message-part.tsx` is updated to assert it lives
in `message-part-render-groups.ts` (the new home) while still
checking that the assistant dispatcher in `message-part.tsx` reads
it. No behavioral change.

message-part.tsx 2103 → 1986 lines (-117). Typecheck GREEN, 478 pass /
3 baseline fail unchanged.
